chore(build): add debug/release aware build utilities

+++ scripts/clean.sh
@@ -1,4 +1,18 @@
 #!/bin/sh
 set -e
 
-sudo rm -rf ./build ./install-test
+BUILD_TYPE="$1"
+
+if [ -z "$BUILD_TYPE" ]; then
+    sudo rm -rf ./build ./install-test
+    exit 0
+fi
+
+if [ "$BUILD_TYPE" != "release" ] && [ "$BUILD_TYPE" != "debug" ]; then
+    echo "usage: $0 [release|debug]"
+    exit 1
+fi
+
+BUILD_DIR="build/$BUILD_TYPE"
+
+sudo rm -rf "$BUILD_DIR" ./install-test
